Separated logic and dom in the main.js file
